+++ server/incore-common/src/main/java/edu/illinois/ncsa/incore/common/auth/Authorizer.java
@@ -10,11 +10,9 @@
 
 package edu.illinois.ncsa.incore.common.auth;
 
-import edu.illinois.ncsa.incore.common.exceptions.IncoreHTTPException;
 import edu.illinois.ncsa.incore.common.models.Space;
 import org.apache.log4j.Logger;
 
-import javax.ws.rs.core.Response;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -32,7 +30,6 @@ public class Authorizer implements IAuthorizer {
 
     public static final String ANONYMOUS_USER = "anonymous";
     private static IAuthorizer instance;
-    private LdapClient ldapClient;
 
     //I don't really like the idea of doing this as a singleton,
     //this thing is going to need some configuration to know what
@@ -47,17 +44,17 @@ public class Authorizer implements IAuthorizer {
 
 
     @Override
-    public Set<PrivilegeLevel> getPrivilegesFor(String user, Privileges spec) {
+    public Set<PrivilegeLevel> getPrivilegesFor(String user, Privileges spec, List<String> userGroups) {
         Set<PrivilegeLevel> privs = getUserSpecificPrivileges(user, spec);
-        privs.addAll(getGroupSpecificPrivileges(user, spec));
+        privs.addAll(getGroupSpecificPrivileges(spec, userGroups));
         return privs;
     }
 
 
     @Override
-    public Set<PrivilegeLevel> getPrivilegesFor(String user, String privilegeSpecJson) {
+    public Set<PrivilegeLevel> getPrivilegesFor(String user, String privilegeSpecJson, List<String> userGroups) {
         Privileges spec = Privileges.fromJson(privilegeSpecJson);
-        return getPrivilegesFor(user, spec);
+        return getPrivilegesFor(user, spec, userGroups);
     }
 
     ////////////////////////////////////////////////////////
@@ -65,8 +62,8 @@ public class Authorizer implements IAuthorizer {
     /////////////////////////////////////////////////////////
 
     @Override
-    public boolean canRead(String user, Privileges privileges) {
-        Set<PrivilegeLevel> privilegesFor = getPrivilegesFor(user, privileges);
+    public boolean canRead(String user, Privileges privileges, List<String> userGroups) {
+        Set<PrivilegeLevel> privilegesFor = getPrivilegesFor(user, privileges, userGroups);
         return (privilegesFor.contains(PrivilegeLevel.READ) ||
             privilegesFor.contains(PrivilegeLevel.WRITE) ||
             privilegesFor.contains(PrivilegeLevel.ADMIN)
@@ -74,8 +71,8 @@ public class Authorizer implements IAuthorizer {
     }
 
     @Override
-    public boolean canRead(String user, String privilegeSpecJson) {
-        Set<PrivilegeLevel> privilegesFor = getPrivilegesFor(user, privilegeSpecJson);
+    public boolean canRead(String user, String privilegeSpecJson, List<String> userGroups) {
+        Set<PrivilegeLevel> privilegesFor = getPrivilegesFor(user, privilegeSpecJson, userGroups);
         return (privilegesFor.contains(PrivilegeLevel.READ) ||
             privilegesFor.contains(PrivilegeLevel.WRITE) ||
             privilegesFor.contains(PrivilegeLevel.ADMIN)
@@ -83,24 +80,24 @@ public class Authorizer implements IAuthorizer {
     }
 
     @Override
-    public boolean canWrite(String user, Privileges privileges) {
-        Set<PrivilegeLevel> privilegesFor = getPrivilegesFor(user, privileges);
+    public boolean canWrite(String user, Privileges privileges, List<String> userGroups) {
+        Set<PrivilegeLevel> privilegesFor = getPrivilegesFor(user, privileges, userGroups);
         return (privilegesFor.contains(PrivilegeLevel.WRITE) ||
             privilegesFor.contains(PrivilegeLevel.ADMIN)
         );
     }
 
     @Override
-    public boolean canWrite(String user, String privilegeSpecJson) {
-        Set<PrivilegeLevel> privilegesFor = getPrivilegesFor(user, privilegeSpecJson);
+    public boolean canWrite(String user, String privilegeSpecJson, List<String> userGroups) {
+        Set<PrivilegeLevel> privilegesFor = getPrivilegesFor(user, privilegeSpecJson, userGroups);
         return (privilegesFor.contains(PrivilegeLevel.WRITE) ||
             privilegesFor.contains(PrivilegeLevel.ADMIN)
         );
     }
 
     @Override
-    public boolean canDelete(String user, Privileges privileges) {
-        Set<PrivilegeLevel> privilegesFor = getPrivilegesFor(user, privileges);
+    public boolean canDelete(String user, Privileges privileges, List<String> userGroups) {
+        Set<PrivilegeLevel> privilegesFor = getPrivilegesFor(user, privileges, userGroups);
         return (privilegesFor.contains(PrivilegeLevel.ADMIN));
     }
 
@@ -113,22 +110,24 @@ public class Authorizer implements IAuthorizer {
      *
      * @param username string name of user
      * @param spaces   a list of all spaces
+     * @param userGroups a list of groups the user belongs to
      * @return a list of accessible spaces
      */
-    public List<Space> getAllSpacesUserCanRead(String username, List<Space> spaces) {
-        return getSpacesUserCanRead(spaces, username);
+    public List<Space> getAllSpacesUserCanRead(String username, List<Space> spaces, List<String> userGroups) {
+        return getSpacesUserCanRead(spaces, username, userGroups);
     }
 
     /**
      * Retrieve all members a user has access to
      *
-     * @param username
+     * @param username string name of user
      * @param spaces   list of all spaces the user has at least READ permission
+     * @param userGroups a list of groups the user belongs to
      * @return a set of all members a user can access with at least READ permission
      */
-    public Set<String> getAllMembersUserHasReadAccessTo(String username, List<Space> spaces) {
+    public Set<String> getAllMembersUserHasReadAccessTo(String username, List<Space> spaces, List<String> userGroups) {
         Set<String> membersSet = new HashSet<>();
-        List<Space> userAccessibleSpaces = getSpacesUserCanRead(spaces, username);
+        List<Space> userAccessibleSpaces = getSpacesUserCanRead(spaces, username, userGroups);
         for (Space space : userAccessibleSpaces) {
             membersSet.addAll(space.getMembers());
         }
@@ -140,10 +139,11 @@ public class Authorizer implements IAuthorizer {
      *
      * @param memberId the String id of the member
      * @param spaces   a list of all spaces
+     * @param userGroups a list of groups the user belongs to
      * @return boolean - true if the user has reading permissions on a space tha contains the member
      */
-    public boolean canUserReadMember(String username, String memberId, List<Space> spaces) {
-        Set<String> userMembersSet = getAllMembersUserHasReadAccessTo(username, spaces);
+    public boolean canUserReadMember(String username, String memberId, List<Space> spaces, List<String> userGroups) {
+        Set<String> userMembersSet = getAllMembersUserHasReadAccessTo(username, spaces, userGroups);
         return userMembersSet.contains(memberId);
     }
 
@@ -154,11 +154,12 @@ public class Authorizer implements IAuthorizer {
      * @param username string name of the user
      * @param memberId the string id of the member
      * @param spaces   a list of all spaces
+     * @param userGroups a list of groups the user belongs to
      * @return boolean - true if the user has admin permissions on a space that contains the member
      */
-    public boolean canUserDeleteMember(String username, String memberId, List<Space> spaces) {
+    public boolean canUserDeleteMember(String username, String memberId, List<Space> spaces, List<String> userGroups) {
         List<Space> userSpaces = spaces.stream()
-            .filter(space -> canUserAccessSpace(space, username, PrivilegeLevel.ADMIN))
+            .filter(space -> canUserAccessSpace(space, username, PrivilegeLevel.ADMIN, userGroups))
             .collect(Collectors.toList());
 
         Set<String> membersSet = new HashSet<>();
@@ -176,11 +177,12 @@ public class Authorizer implements IAuthorizer {
      * @param username string name of the user
      * @param memberId string id of the member
      * @param spaces   a list of all spaces
+     * @param userGroups a list of groups the user belongs to
      * @return boolean - true if the user has write or admin permissions on a space that contains the member
      */
-    public boolean canUserWriteMember(String username, String memberId, List<Space> spaces) {
+    public boolean canUserWriteMember(String username, String memberId, List<Space> spaces, List<String> userGroups) {
         List<Space> userSpaces = spaces.stream()
-            .filter(space -> canUserAccessSpace(space, username, PrivilegeLevel.WRITE))
+            .filter(space -> canUserAccessSpace(space, username, PrivilegeLevel.WRITE, userGroups))
             .collect(Collectors.toList());
 
         Set<String> membersSet = new HashSet<>();
@@ -193,31 +195,23 @@ public class Authorizer implements IAuthorizer {
     /**
      * This method determines if a user is an admin or not.
      *
-     * @param username string name of the user
+     * @param userGroups list of strings of groups user belongs to
      * @return boolean - true if the user is admin
      */
-    public boolean isUserAdmin(String username) {
-        try {
-            LdapClient ldapClient = getLdapClient();
-            Set<String> userGroups = ldapClient.getUserGroups(username);
-
-            return userGroups.contains(System.getenv("AUTH_LDAP_ADMINS"));
-        } catch (Exception e) {
-            logger.error(e);
-            throw new IncoreHTTPException(Response.Status.INTERNAL_SERVER_ERROR, "Internal server error.");
-        }
+    public boolean isUserAdmin(List<String> userGroups) {
+        return (userGroups.contains("incore_admin") || userGroups.contains("incore_ncsa"));
     }
 
     ////////////////////////////////////////////////////////
     // private methods
     /////////////////////////////////////////////////////////
 
-    private Set<PrivilegeLevel> getGroupSpecificPrivileges(String user, Privileges spec) {
+    private Set<PrivilegeLevel> getGroupSpecificPrivileges(Privileges spec, List<String> userGroups) {
         if (spec == null) {
-            return allowThisUser(user);
+            return allowThisUser();
         }
         try {
-            if (isUserAdmin(user)) {
+            if (isUserAdmin(userGroups)) {
                 HashSet<PrivilegeLevel> admin = new HashSet<>();
                 admin.add(PrivilegeLevel.ADMIN);
                 admin.add(PrivilegeLevel.WRITE);
@@ -225,14 +219,13 @@ public class Authorizer implements IAuthorizer {
                 return admin;
             }
 
-            Set<String> userGroups = ldapClient.getUserGroups(user);
             Set<PrivilegeLevel> privs = spec.groupPrivileges.keySet().stream()
                 .filter(userGroups::contains)
                 .map(key -> spec.groupPrivileges.get(key))
                 .collect(Collectors.toSet());
 
             //if the user is in the magic view-all group, give them read access
-            if (userGroups.contains(System.getenv("AUTH_LDAP_VIEW_ALL"))) {
+            if (userGroups.contains("incore_admin") || userGroups.contains("incore_ncsa")) {
                 privs.add(PrivilegeLevel.READ);
             }
 
@@ -245,15 +238,15 @@ public class Authorizer implements IAuthorizer {
         return new HashSet<PrivilegeLevel>();
     }
 
-    private Set<PrivilegeLevel> allowThisUser(String user) {
-        Set<PrivilegeLevel> allower = new HashSet<>();
-        allower.add(PrivilegeLevel.ADMIN);
-        return allower;
+    private Set<PrivilegeLevel> allowThisUser() {
+        Set<PrivilegeLevel> allowed = new HashSet<>();
+        allowed.add(PrivilegeLevel.ADMIN);
+        return allowed;
     }
 
     private Set<PrivilegeLevel> getUserSpecificPrivileges(String user, Privileges spec) {
         if (spec == null) {
-            return allowThisUser(user);
+            return allowThisUser();
         }
         try {
             return spec.userPrivileges.keySet().stream()
@@ -266,23 +259,17 @@ public class Authorizer implements IAuthorizer {
         return new HashSet<PrivilegeLevel>();
     }
 
-    private LdapClient getLdapClient() {
-        if (ldapClient == null) {
-            ldapClient = new LdapClient();
-        }
-        return ldapClient;
-    }
-
     /**
      * Retrieve all spaces that a user has at least READ permissions
      *
      * @param spaces   List of all spaces in spacedb
      * @param username Name of user's space
+     * @param userGroups list of strings of groups user belongs to
      * @return List of spaces that a user can access with at least READ permission
      */
-    private List<Space> getSpacesUserCanRead(List<Space> spaces, String username) {
+    private List<Space> getSpacesUserCanRead(List<Space> spaces, String username, List<String> userGroups) {
         return spaces.stream()
-            .filter(space -> canUserAccessSpace(space, username, PrivilegeLevel.READ))
+            .filter(space -> canUserAccessSpace(space, username, PrivilegeLevel.READ, userGroups))
             .collect(Collectors.toList());
     }
 
@@ -294,22 +281,23 @@ public class Authorizer implements IAuthorizer {
      * @param space          the space retrieved from the repository
      * @param username       the string name of the user
      * @param privilegeLevel the privilege level we want to check for
+     * @param userGroups list of strings of groups user belongs to
      * @return boolean - true if the user has the specified privilege level
      */
-    private boolean canUserAccessSpace(Space space, String username, PrivilegeLevel privilegeLevel) {
+    private boolean canUserAccessSpace(Space space, String username, PrivilegeLevel privilegeLevel, List<String> userGroups) {
         if (privilegeLevel == PrivilegeLevel.READ) {
             return space.getUserPrivilegeLevel(username) != null ||
                 space.getGroupPrivilegeLevel(username) != null ||
-                !getGroupSpecificPrivileges(username, space.getPrivileges()).isEmpty();
+                !getGroupSpecificPrivileges(space.getPrivileges(), userGroups).isEmpty();
         } else if (privilegeLevel == PrivilegeLevel.WRITE) {
             return space.getUserPrivilegeLevel(username) == PrivilegeLevel.WRITE || space.getUserPrivilegeLevel(username) == PrivilegeLevel.ADMIN ||
                 space.getGroupPrivilegeLevel(username) == PrivilegeLevel.WRITE || space.getGroupPrivilegeLevel(username) == PrivilegeLevel.ADMIN ||
-                (getGroupSpecificPrivileges(username, space.getPrivileges())).contains(PrivilegeLevel.WRITE) ||
-                (getGroupSpecificPrivileges(username, space.getPrivileges())).contains(PrivilegeLevel.ADMIN);
+                (getGroupSpecificPrivileges(space.getPrivileges(), userGroups)).contains(PrivilegeLevel.WRITE) ||
+                (getGroupSpecificPrivileges(space.getPrivileges(), userGroups)).contains(PrivilegeLevel.ADMIN);
         } else if (privilegeLevel == PrivilegeLevel.ADMIN) {
             return space.getUserPrivilegeLevel(username) == PrivilegeLevel.ADMIN ||
                 space.getGroupPrivilegeLevel(username) == PrivilegeLevel.ADMIN ||
-                (getGroupSpecificPrivileges(username, space.getPrivileges())).contains(PrivilegeLevel.ADMIN);
+                (getGroupSpecificPrivileges(space.getPrivileges(), userGroups)).contains(PrivilegeLevel.ADMIN);
         }
         return false;
     }