+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Query parameters for sorting by given field in ascending and descending order [#111](https://github.com/IN-CORE/incore-services/issues/111)
 - Get user group function to read and parse user groups from headers [#120](https://github.com/IN-CORE/incore-services/issues/120)
-- 
+
+### Changed
+- Remove LdapClient and refactor the Authorizer class along with its interface to take in userGroups parameter [#118](https://github.com/IN-CORE/incore-services/issues/118)
+
 ## [1.13.1] - 2023-03-15
 ### Fixed
 - Earthquake hazard raster didn't handle values below the threshold that were set to null[#108](https://github.com/IN-CORE/incore-services/issues/108)